separate preview tools; inspector partially working;
