+++ assets/src/apps/delivery/components/PreviewTools.tsx
@@ -1,14 +1,12 @@
 /* eslint-disable no-prototype-builtins */
 /* eslint-disable react/prop-types */
-import { defaultGlobalEnv, getEnvState } from '../../../adaptivity/scripting';
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { v4 as uuidv4 } from 'uuid';
 import { selectCurrentActivity } from '../store/features/activities/slice';
-import { triggerCheck } from '../store/features/adaptivity/actions/triggerCheck';
-import { setLastCheckResults } from '../store/features/adaptivity/slice';
 import { navigateToActivity } from '../store/features/groups/actions/deck';
-import { CapiVariableTypes, getCapiType } from '../../../adaptivity/capi';
+import Adaptivity from './preview-tools/Adaptivity';
+import Inspector from './preview-tools/Inspector';
+import ScreenSelector from './preview-tools/ScreenSelector';
 
 // Title Component
 interface TitleProps {
@@ -40,579 +38,6 @@ const Title: React.FC<any> = (props: TitleProps) => {
         {title === 'Adaptivity' && <AdaptivityIcon />}
         {title === 'Inspector' && <InspectorIcon />}
         <div className="pt-title">{title}</div>
-      </div>
-    </div>
-  );
-};
-
-// Screen Selector View
-interface ScreenSelectorProps {
-  sequence: any;
-  navigate: any;
-  currentActivity: any;
-}
-const ScreenSelector: React.FC<ScreenSelectorProps> = ({
-  sequence,
-  navigate,
-  currentActivity,
-}: ScreenSelectorProps) => {
-  return (
-    <div className={`preview-tools-view`}>
-      <ol className="list-group list-group-flush">
-        {sequence?.map((s: any, i: number) => {
-          return (
-            <li
-              key={i}
-              className={`list-group-item pl-5 py-1 list-group-item-action${
-                currentActivity?.id === s.sequenceId ? ' active' : ''
-              }`}
-            >
-              <a
-                href=""
-                className={currentActivity?.id === s.sequenceId ? 'selected' : ''}
-                onClick={(e) => {
-                  e.preventDefault();
-                  navigate(s.sequenceId);
-                }}
-              >
-                {s.sequenceName}
-              </a>
-            </li>
-          );
-        })}
-      </ol>
-    </div>
-  );
-};
-
-interface AdaptivityProps {
-  currentActivity: any;
-}
-// Adaptivity Placeholder
-const Adaptivity: React.FC<AdaptivityProps> = ({ currentActivity }) => {
-  const [expandedRules, setExpandedRules]: any = useState({});
-  const dispatch = useDispatch();
-
-  const triggerAction = async (e: any, rule: any) => {
-    e.preventDefault();
-
-    // Strip out any rules that are not the selected rule
-    const customRules = currentActivity.authoring?.rules?.filter(
-      (activityRule: any) => activityRule.id === rule.id,
-    );
-    // Construct custom check results to fire after triggering the check event
-    const customCheckResults = [
-      {
-        params: {
-          actions: [...rule.event.params.actions],
-          correct: !!rule.correct,
-          default: !!rule.default,
-          order: rule.priority,
-        },
-        type: rule.event.type,
-      },
-    ];
-
-    // Send in the custom rules for processing
-    dispatch(triggerCheck({ activityId: currentActivity.id, customRules }));
-    // Fire off our custom check results
-    await dispatch(setLastCheckResults({ results: customCheckResults }));
-  };
-
-  useEffect(() => {
-    setExpandedRules({});
-  }, [currentActivity]);
-
-  // helper because of lint
-  const hasOwnProperty = (obj: any, property: string) =>
-    Object.prototype.hasOwnProperty.call(obj || {}, property);
-
-  return (
-    <div className="adaptivity">
-      <div className="accordion">
-        {/* InitState */}
-        {currentActivity?.content?.custom?.facts?.length > 0 && (
-          <div key={`init-${currentActivity.id}`} className="card initState">
-            <div className="card-header p-2" id={`initHeading-${currentActivity.id}`}>
-              <h2 className="mb-0">
-                <button
-                  className="btn btn-link btn-block text-left"
-                  type="button"
-                  data-toggle="collapse"
-                  data-target={`#collapse`}
-                  aria-expanded={expandedRules[`initState-${currentActivity.id}`]}
-                  aria-controls={`collapse`}
-                  onClick={(e) =>
-                    setExpandedRules({
-                      ...expandedRules,
-                      [`initState-${currentActivity.id}`]: !expandedRules[
-                        `initState-${currentActivity.id}`
-                      ],
-                    })
-                  }
-                >
-                  <span
-                    className={`chevron-arrow mr-2${
-                      expandedRules[`initState-${currentActivity.id}`] ? ' rotate' : ''
-                    }`}
-                  >
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      fill="currentColor"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M4.646 1.646a.5.5 0 01.708 0l6 6a.5.5 0 010 .708l-6 6a.5.5 0 01-.708-.708L10.293 8 4.646 2.354a.5.5 0 010-.708z"
-                      />
-                    </svg>
-                  </span>
-                  Initial State
-                </button>
-              </h2>
-            </div>
-            <div
-              id={`collapse`}
-              className="collapse"
-              aria-labelledby={`initHeading-${currentActivity.id}`}
-            >
-              <div className="mt-2 pt-2 px-3 font-weight-bold text-uppercase">Facts</div>
-              <div className="card-body pt-2">
-                <ul className="list-group">
-                  {currentActivity?.content?.custom?.facts.map((fact: any, factIndex: number) => (
-                    <li key={factIndex} className="list-group-item">{`${fact.target} ${
-                      fact.operator
-                    } ${JSON.stringify(fact.value)}`}</li>
-                  ))}
-                </ul>
-              </div>
-            </div>
-          </div>
-        )}
-        {/* Rules */}
-        {currentActivity?.authoring?.rules?.map((rule: any, ruleIndex: number) => (
-          <div
-            key={`${rule.id}-${ruleIndex}`}
-            className={`card${rule.correct ? ' correct' : ' incorrect'}`}
-          >
-            <div className="card-header p-2" id={`heading${ruleIndex}`}>
-              <h2 className="mb-0">
-                <button
-                  className="btn btn-link btn-block text-left"
-                  type="button"
-                  data-toggle="collapse"
-                  data-target={`#collapse${ruleIndex}`}
-                  aria-expanded={expandedRules[`rule-${ruleIndex}`]}
-                  aria-controls={`collapse${ruleIndex}`}
-                  onClick={(e) =>
-                    setExpandedRules({
-                      ...expandedRules,
-                      [`rule-${ruleIndex}`]: !expandedRules[`rule-${ruleIndex}`],
-                    })
-                  }
-                >
-                  <span
-                    className={`chevron-arrow mr-2${
-                      expandedRules[`rule-${ruleIndex}`] ? ' rotate' : ''
-                    }`}
-                  >
-                    <svg
-                      xmlns="http://www.w3.org/2000/svg"
-                      width="16"
-                      height="16"
-                      fill="currentColor"
-                    >
-                      <path
-                        fillRule="evenodd"
-                        d="M4.646 1.646a.5.5 0 01.708 0l6 6a.5.5 0 010 .708l-6 6a.5.5 0 01-.708-.708L10.293 8 4.646 2.354a.5.5 0 010-.708z"
-                      />
-                    </svg>
-                  </span>
-                  {rule.name}
-                </button>
-              </h2>
-            </div>
-            <div
-              id={`collapse${ruleIndex}`}
-              className="collapse"
-              aria-labelledby={`heading${ruleIndex}`}
-            >
-              <div className="mt-2 pt-2 px-3 font-weight-bold text-uppercase">Conditions</div>
-              <div className="card-body pt-2">
-                <ul className="list-group">
-                  {/* Conditions for ALL */}
-                  {hasOwnProperty(rule.conditions, 'all') && (
-                    <>
-                      {rule.conditions?.all?.length === 0 && (
-                        <li className="list-group-item">
-                          No conditions. This rule will always fire.
-                        </li>
-                      )}
-                      {rule.conditions?.all?.length > 0 &&
-                        rule.conditions.all.map((condition: any, conditionIndex: number) => (
-                          <li key={conditionIndex} className="list-group-item">
-                            {`${condition.fact} ${condition.operator} ${JSON.stringify(
-                              condition.value,
-                            )}`}
-                          </li>
-                        ))}
-                    </>
-                  )}
-                  {/* Conditions for ANY */}
-                  {hasOwnProperty(rule.conditions, 'any') && (
-                    <>
-                      {rule.conditions?.any?.length === 0 && (
-                        <li className="list-group-item">
-                          No conditions. This rule will always fire.
-                        </li>
-                      )}
-                      {rule.conditions?.any?.length > 0 &&
-                        rule.conditions.any.map((condition: any, conditionIndex: number) => (
-                          <li key={conditionIndex} className="list-group-item">
-                            {`${condition.fact} ${condition.operator} ${JSON.stringify(
-                              condition.value,
-                            )}`}
-                          </li>
-                        ))}
-                    </>
-                  )}
-                </ul>
-              </div>
-              <div className="d-flex justify-content-between align-items-center px-3 font-weight-bold text-uppercase">
-                Actions{' '}
-                <button
-                  onClick={(e) => triggerAction(e, rule)}
-                  type="button"
-                  className="btn btn-sm btn-outline-primary d-flex px-1"
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    width="16"
-                    height="16"
-                    fill="currentColor"
-                  >
-                    <path d="M11.596 8.697l-6.363 3.692c-.54.313-1.233-.066-1.233-.697V4.308c0-.63.692-1.01 1.233-.696l6.363 3.692a.802.802 0 010 1.393z" />
-                  </svg>
-                </button>
-              </div>
-              <div className="card-body pt-2">
-                <ul className="list-group">
-                  {rule.event?.params?.actions?.length === 0 && (
-                    <li className="list-group-item">No actions assigned.</li>
-                  )}
-                  {rule.event?.params?.actions?.length > 0 &&
-                    rule.event.params.actions.map((action: any, actionIndex: number) => (
-                      <li key={actionIndex} className="list-group-item">
-                        <span className="text-capitalize">{action.type}</span>
-                      </li>
-                    ))}
-                </ul>
-              </div>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-interface InspectorProps {
-  currentActivity: any;
-}
-// Inspector Placeholder
-const Inspector: React.FC<InspectorProps> = ({ currentActivity }) => {
-  const [globalState, setGlobalState] = useState<any>({});
-  const [globalInputState, setGlobalInputState] = useState<any>({});
-  const [sessionState, setSessionState] = useState<any>({});
-  const [stageState, setStageState] = useState<any>({});
-
-  const unflatten = (data: any) => {
-    // https://stackoverflow.com/questions/42694980/how-to-unflatten-a-javascript-object-in-a-daisy-chain-dot-notation-into-an-objec
-    const result = {};
-    for (const i in data) {
-      const keys = i.split('.');
-      keys.reduce(function (r: any, e: any, j) {
-        return (
-          r[e] || (r[e] = isNaN(Number(keys[j + 1])) ? (keys.length - 1 == j ? data[i] : {}) : [])
-        );
-      }, result);
-    }
-    return result;
-  };
-  const isArray = (array: any) => {
-    return !!array && array.constructor === Array;
-  };
-  const isObject = (object: any) => {
-    return !!object && object.constructor === Object;
-  };
-  const hasNesting: any = (thing: any) => {
-    if (isObject(thing) && Object.keys(thing).length > 0) {
-      return true;
-    }
-    if (isArray(thing) && thing.length > 0) {
-      return true;
-    }
-    return false;
-  };
-
-  const getSessionState = (): any => {
-    const statePuff: any = unflatten(globalState);
-    return setSessionState({ ...statePuff['session'] });
-  };
-
-  const getStageState = (): any => {
-    const statePuff: any = unflatten(globalState);
-    return setStageState({ ...statePuff[`${currentActivity.id}|stage`] });
-  };
-
-  useEffect(() => {
-    setStageState({});
-    setSessionState({});
-    setTimeout(() => {
-      // TODO : figure out a better way to setGlobalState AFTER state reset to fix timing jank
-      setGlobalState(getEnvState(defaultGlobalEnv));
-    }, 50);
-  }, [currentActivity]);
-
-  useEffect(() => {
-    getSessionState();
-    getStageState();
-  }, [globalState]);
-
-  interface AutoDetectInputProps {
-    label: string;
-    value: any;
-    state?: any;
-    // onChange: ()=>void;
-  }
-  const AutoDetectInput: React.FC<AutoDetectInputProps> = ({ label, value, state }): any => {
-    console.log('🚀 > file: PreviewTools.tsx > line 390 > { label, value ,state}', {
-      label,
-      value,
-      state,
-    });
-    const uuid = uuidv4();
-    const capiType = getCapiType(value);
-    switch (capiType) {
-      case CapiVariableTypes.BOOLEAN:
-        return (
-          <div className="custom-control custom-switch">
-            <input
-              type="checkbox"
-              className="custom-control-input"
-              id={uuid}
-              defaultChecked={value}
-              // onChange={(e) => setInputState({ ...inputState, value: e.target.value })}
-            />
-            <label className="custom-control-label" htmlFor={uuid}></label>
-          </div>
-        );
-
-      case CapiVariableTypes.ENUM:
-        return (
-          // TODO : wire this up
-          <div className="user-input">
-            <span className="stateKey" title="session.visits.q:1541198781354:733">
-              q:1541198781354:733
-            </span>
-            {/* Dropdown example */}
-            <select className="custom-select custom-select-sm" defaultValue="3">
-              <option value="1">One</option>
-              <option value="2">Two</option>
-              <option value="3">Three</option>
-              <option value="4">
-                This option has a very long text node that may stretch out the drop down. What
-                happens?
-              </option>
-            </select>
-          </div>
-        );
-
-      default:
-        return (
-          <input
-            type="text"
-            className="input-group-sm stateValue"
-            aria-label={label}
-            defaultValue={JSON.stringify(value)}
-          />
-        );
-    }
-  };
-
-  interface NestedStateDisplayProps {
-    rootLevel: any;
-    levelIndex: number;
-    state: any;
-  }
-  const NestedStateDisplay: React.FC<NestedStateDisplayProps> = ({
-    rootLevel,
-    levelIndex,
-    state,
-  }) => {
-    const [expandedPanels, setExpandedPanels]: any = useState({});
-    return (
-      <li key={`leaf-branch-${rootLevel}${levelIndex}`} className="list-group-item is-parent">
-        {/* TODO Toggle even / odd based on index */}
-        <div
-          className="card-header p-0 m-0 rounded-lg mt-2 even"
-          id={`heading${rootLevel}${levelIndex}`}
-        >
-          <button
-            className="btn btn-link btn-block text-left"
-            type="button"
-            // TODO: figure out why Bootstrap collapse is breaking in recursion
-            // data-toggle="collapse"
-            // data-target={`#collapse${rootLevel}${levelIndex}`}
-            aria-expanded={expandedPanels[`panel-${rootLevel}${levelIndex}`]}
-            aria-controls={`collapse${rootLevel}${levelIndex}`}
-            onClick={() =>
-              setExpandedPanels({
-                ...expandedPanels,
-                [`panel-${rootLevel}${levelIndex}`]: !expandedPanels[
-                  `panel-${rootLevel}${levelIndex}`
-                ],
-              })
-            }
-          >
-            <span
-              className={`chevron-arrow mr-2${
-                expandedPanels[`panel-${rootLevel}${levelIndex}`] ? ' rotate' : ''
-              }`}
-            >
-              <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor">
-                <path
-                  fillRule="evenodd"
-                  d="M4.646 1.646a.5.5 0 01.708 0l6 6a.5.5 0 010 .708l-6 6a.5.5 0 01-.708-.708L10.293 8 4.646 2.354a.5.5 0 010-.708z"
-                />
-              </svg>
-            </span>
-            {rootLevel}
-          </button>
-        </div>
-        <div
-          id={`collapse${rootLevel}${levelIndex}`}
-          // TODO: reset className to 'collapse' after figuring out Bootstrap recursion issue
-          className={`${
-            expandedPanels[`panel-${rootLevel}${levelIndex}`] ? '' : 'visually-hidden'
-          }`}
-          aria-labelledby={`heading${rootLevel}${levelIndex}`}
-        >
-          <ul className="list-group list-group-flush">
-            {state[rootLevel] &&
-              Object.keys(state[rootLevel])
-                .sort()
-                .map((level2: any) =>
-                  !hasNesting(state[rootLevel][level2]) ? (
-                    <li key={`flat-${level2}${levelIndex + 1}`} className="list-group-item pr-0">
-                      <div className="user-input">
-                        <span className="stateKey" title={level2}>
-                          {level2}
-                        </span>
-                        <AutoDetectInput
-                          label={level2}
-                          value={state[rootLevel][level2]}
-                          state={{ ...state[rootLevel] }}
-                        />
-                      </div>
-                    </li>
-                  ) : (
-                    <NestedStateDisplay
-                      key={`${level2}${levelIndex + 1}`}
-                      rootLevel={level2}
-                      levelIndex={levelIndex + 1}
-                      state={{ [level2]: { ...state[rootLevel][level2] } }}
-                    />
-                  ),
-                )}
-          </ul>
-        </div>
-      </li>
-    );
-  };
-
-  interface StateDisplayProps {
-    label: string;
-    state: any;
-  }
-  const StateDisplay: React.FC<StateDisplayProps> = ({ label, state }) => {
-    const [expandedPanels, setExpandedPanels]: any = useState({});
-
-    return Object.keys(state).length > 0 ? (
-      <div className="card even">
-        <div className="card-header p-2" id={`headingRoot${label}`}>
-          <h2 className="mb-0">
-            <button
-              className="btn btn-link btn-block text-left"
-              type="button"
-              // data-toggle="collapse"
-              // data-target={`#collapseRoot${label}`}
-              aria-expanded={expandedPanels[`panel-Root${label}`]}
-              aria-controls={`collapseRoot${label}`}
-              onClick={() =>
-                setExpandedPanels({
-                  ...expandedPanels,
-                  [`panel-Root${label}`]: !expandedPanels[`panel-Root${label}`],
-                })
-              }
-            >
-              <span
-                className={`chevron-arrow mr-2${
-                  expandedPanels[`panel-Root${label}`] ? ' rotate' : ''
-                }`}
-              >
-                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor">
-                  <path
-                    fillRule="evenodd"
-                    d="M4.646 1.646a.5.5 0 01.708 0l6 6a.5.5 0 010 .708l-6 6a.5.5 0 01-.708-.708L10.293 8 4.646 2.354a.5.5 0 010-.708z"
-                  />
-                </svg>
-              </span>
-              {label}
-            </button>
-          </h2>
-        </div>
-        <div
-          id={`collapseRoot${label}`}
-          className={`${expandedPanels[`panel-Root${label}`] ? '' : 'visually-hidden'}`}
-          aria-labelledby={`headingRoot${label}`}
-        >
-          <div className="card-body py-2">
-            <ul className="list-group list-group-flush">
-              {Object.keys(state)
-                .sort()
-                .map((level1: any, index: number) =>
-                  !hasNesting(state[level1]) ? (
-                    <li key={`leaf-${level1}${index}`} className="list-group-item pr-0">
-                      <div className="user-input">
-                        <span className="stateKey" title={level1}>
-                          {level1}
-                        </span>
-                        <AutoDetectInput label={level1} value={state[level1]} state={state} />
-                      </div>
-                    </li>
-                  ) : (
-                    <NestedStateDisplay
-                      key={`${level1}${index}`}
-                      rootLevel={level1}
-                      levelIndex={1}
-                      state={state}
-                    />
-                  ),
-                )}
-            </ul>
-          </div>
-        </div>
-      </div>
-    ) : null;
-  };
-
-  return (
-    <div className="inspector">
-      <div className="accordion">
-        <StateDisplay label="Session" state={sessionState} />
-        <StateDisplay label="Stage" state={stageState} />
       </div>
     </div>
   );
